Move ColorMode enum to Helper/Enum/ subfolder per convention.
